refactor: move patterns.ts to /constants
