+++ .env.example
@@ -1,7 +1,7 @@
 VITE_APP_PROJECT_ID="your-project-id"
 VITE_IPFS_KEY="your-ipfs-key"
 VITE_DIAMOND_ADDRESS="0x6Bf26Ec92291123Be0f56AB1A43D93837E9b880a"
-VITE_PREFERRED_NETWORK_ID=5
+VITE_PREFERRED_NETWORK_ID=80001
 VITE_VERIFICATION_API_URL="http://localhost:43210/api"
 VITE_SEARCHSECO_API_URL="http://localhost:25566/api"
 VITE_VERIFY_CONTRACT="0x..."
