+++ src/lib/constants/patterns.ts
